refactor(provider): folder of providers
